Header naming and integration fixes + working on sign up page + pg upgrade
