+++ frontend/package.json
@@ -19,7 +19,7 @@
     "neo4j-driver": "^5.28.1",
     "openai": "^4.89.0",
     "path-browserify": "^1.0.1",
-    "pg": "^8.13.3",
+    "pg": "^8.14.1",
     "postcss-loader": "^8.1.1",
     "querystring-es3": "^0.2.1",
     "react": "^19.0.0",
